requirements.txt
certifi
Faker
google-generativeai
python-dotenv
openai
pandas